Create ReactiveSwift and ReactiveSwiftTests target
- Adjust import statement based on compiled target
- Extract `DynamicProperty` to separate file and add it to mac targets
