+++ ReactiveCocoaTests/Swift/ActionSpec.swift
@@ -9,7 +9,12 @@
 import Result
 import Nimble
 import Quick
+
+#if REACTIVE_SWIFT
+import ReactiveSwift
+#else
 import ReactiveCocoa
+#endif
 
 class ActionSpec: QuickSpec {
 	override func spec() {
