+++ ReactiveCocoaTests/Swift/TestLogger.swift
@@ -7,7 +7,13 @@
 //
 
 import Foundation
-@testable import ReactiveCocoa
+@testable
+
+#if REACTIVE_SWIFT
+import ReactiveSwift
+#else
+import ReactiveCocoa
+#endif
 
 final class TestLogger {
 	private var expectations: [String -> Void]
